week7,8 코드 review

+++ src/main/java/org/cookieandkakao/babting/domain/calendar/entity/Reminder.java
@@ -22,17 +22,17 @@ public class Reminder {
     private Event event;
 
     @Column
-    private Long remindTime;
+    private Integer remindTime;
 
     protected Reminder() {
     }
 
-    public Reminder(Event event, Long remindTime) {
+    public Reminder(Event event, Integer remindTime) {
         this.event = event;
         this.remindTime = remindTime;
     }
 
-    public Long getRemindTime() {
+    public Integer getRemindTime() {
         return remindTime;
     }
